# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "version": "1.0.0",
      "title": "Plainx middleware api documentation",
      "description": "API documentation for plainx middleware",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "https://plainx-midddleware-engine.azurewebsites.net",
    "basePath": "/",
    "components": {
        "securitySchemes": {
          "jwt": {
            "type": "http",
            "scheme": "bearer",
            "name": "x-access-token",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [{
        "jwt": []
      }],
...
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/project": {
        "post": {
            "tags": [
              "Projects"
            ],
            "parameters": [
              {
                "name": "project",
                "in": "body",
                "description": "projects that we want to create",
                "schema": {
                  "$ref": "#/definitions/Project"
                }
              }
            ],
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/definitions/Project"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "New Project is created",
                "schema": {
                  "$ref": "#/definitions/Project"
                }
              },
              "401": {
                  "description": "Email address must be unique",
                  "schema": {
                    "$ref": "#/definitions/Project"
                  }
              }
            }
        }
      },
      "/project/all/{userId}": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get all projects in system",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Projects"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
      "/project/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get project with given ID",
          "responses": {
            "200": {
              "description": "Project is found",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
      "/project/model_comparison/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get model comparison details of project with given projectId",
          "responses": {
            "200": {
              "description": "Model comparison details found successfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem finding the data. Please try again after sometime"
            }
          }
        }
      },
      "/project/data_summary/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get model summary details of project with given projectId",
          "responses": {
            "200": {
              "description": "Model details found",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem finding the data. Please try again after sometime"
            }
          }
        }
      },
      "/project/bias_results/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get bias results of project with given projectId",
          "responses": {
            "200": {
              "description": "Bias results found sucessfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem finding the data. Please try again after sometime"
            }
          }
        }
      },
      "/project/bias_features/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get bias features of project with given projectId",
          "responses": {
            "200": {
              "description": "Bias features found sucessfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem finding the data. Please try again after sometime"
            }
          }
        }
      },
      "/project/eda_data/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get eda data of project with given projectId",
          "responses": {
            "200": {
              "description": "EDA data found sucessfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem finding the data. Please try again after sometime"
            }
          }
        }
      },
      "/project/business_tree/{projectId}": {
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we need to fetch",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Projects"
          ],
          "summary": "Get binary tree with given ID",
          "responses": {
            "200": {
              "description": "Binary Tree data found succesfully",
              "schema": {
                "$ref": "#/definitions/Project"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            },
            "404":{
              "description": "No project found with given id"
            },
            "500":{
              "description":"There was a problem fetching data."
            }
          }
        },
        "put": {
          "summary": "Update binary tree with given ID",
          "tags": [
            "Projects"
          ],
          "parameters": [
            {
              "name": "Project",
              "in": "body",
              "description": "Binary Tree with new values of properties",
              "schema": {
                "$ref": "#/definitions/BusinessTree"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Binary Tree is updated",
              "schema": {
                "$ref": "#/definitions/BusinessTree"
              }
            },
            "404": {
              "description": "No project found with given id"
            },
            "500":
            {
              "description": "There was a problem updating the Binary tree data"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Project": {
        "properties": {
          "userId": { "type": "string"},
          "projectName": { "type": "string" },
          "projectStage": { "type": "string" },
          "industry" : { "type": "string" },
          "useCase": { "type": "string" },
          "model": {"type": "object", "$ref": "#/definitions/Model"},
          "datasetFile": { "type": "string" },
          "datasetSize": { "type": "string" },
          "datasetName": { "type": "string" }
        },
        "required": [
            "userId"
        ]
      },
      "Model": {
        "properties": {
          "modelName": { "type": "string"},
          "modelType": { "type": "string"},
          "modelFile": { "type": "string"},
          "dataToPredict": { "type": "string"}
        }
      },
      "Projects": {
          "type": "array",
          "$ref": "#/definitions/Project"
      },
...
        "TreeGraphDataAttributes":{
          "properties": {
           " name":{"type":"string"},
            "isSelected":{"type":"boolean"},
           " filterKey":{"type":"string"},
           " filterCondition":{"type":"string"},
            "filterCriteria":{"type":"string"}
          }
        },
        "NodeDisplayData":{
          "properties":{
            "bias":{"type":"string"},
            "probability":{"type":"number"}
          },
          "required": [
            "projectId"
        ]
        },
      "BinaryNode":{
        "properties":{
          "name":{"type":"string"},
         "attributes":{
          "type":"object",
          "$ref":"#/definitions/TreeGraphDataAttributes"
        },
         "nodeDisplayData":{
          "type":"object",
          "$ref":"#/definitions/NodeDisplayData"
        }
        }
      },
      "Children":{
        "type":"object",
        "$ref":"#/definitions/BinaryNode"
      },
      "BusinessTree":{
        "properties":{
          "name":{"type":"string"},
         "attributes":{
          "type":"object",
          "$ref":"#/definitions/TreeGraphDataAttributes"
        },
         "nodeDisplayData":{
          "type":"object",
          "$ref":"#/definitions/NodeDisplayData"
        },
         "children":{
          "type":"array",
          "items":{"$ref":"#/definitions/Children"}
        }
        }
      }
    }
  }